fix issues with variables expansions

+++ tests-ng/dotfiles-dyn-paths.sh
@@ -1,13 +1,14 @@
 #!/usr/bin/env bash
 # author: deadc0de6 (https://github.com/deadc0de6)
-# Copyright (c) 2017, deadc0de6
+# Copyright (c) 2019, deadc0de6
 #
-# test variables from yaml file
+# test dotfile dynamic paths
 # returns 1 in case of error
 #
 
 # exit on first error
 set -e
+#set -v
 
 # all this crap to get current path
 rl="readlink -f"
@@ -48,10 +49,10 @@ echo -e "\e[96m\e[1m==> RUNNING $(basename $BASH_SOURCE) <==\e[0m"
 # the dotfile source
 tmps=`mktemp -d --suffix='-dotdrop-tests'`
 mkdir -p ${tmps}/dotfiles
-#echo "dotfile source: ${tmps}"
+echo "dotfiles source (dotpath): ${tmps}"
 # the dotfile destination
 tmpd=`mktemp -d --suffix='-dotdrop-tests'`
-#echo "dotfile destination: ${tmpd}"
+echo "dotfiles destination: ${tmpd}"
 
 # create the config file
 cfg="${tmps}/config.yaml"
@@ -62,50 +63,33 @@ config:
   create: true
   dotpath: dotfiles
 variables:
-  var1: "this is some test"
-  var2: 12
-  var3: another test
+  dst: "${tmpd}/abc"
+dynvariables:
+  src: "echo abc"
 dotfiles:
   f_abc:
-    dst: ${tmpd}/abc
-    src: abc
+    dst: "{{@@ dst @@}}"
+    src: "{{@@ src @@}}"
 profiles:
   p1:
     dotfiles:
     - f_abc
-    variables:
-      var1: "this is some sub-test"
-  p2:
-    include:
-    - p1
-    variables:
-      var2: 42
 _EOF
 #cat ${cfg}
 
-# create the dotfile
-echo "{{@@ var1 @@}}" > ${tmps}/dotfiles/abc
-echo "{{@@ var2 @@}}" >> ${tmps}/dotfiles/abc
-echo "{{@@ var3 @@}}" >> ${tmps}/dotfiles/abc
-echo "test" >> ${tmps}/dotfiles/abc
+# create the dotfiles
+echo "abc" > ${tmps}/dotfiles/abc
+
+###########################
+# test install and compare
+###########################
 
 # install
-cd ${ddpath} | ${bin} install -f -c ${cfg} -p p1
+cd ${ddpath} | ${bin} install -f -c ${cfg} -p p1 -b -V
+[ "$?" != "0" ] && exit 1
 
-cat ${tmpd}/abc
-grep '^this is some sub-test' ${tmpd}/abc >/dev/null
-grep '^12' ${tmpd}/abc >/dev/null
-grep '^another test' ${tmpd}/abc >/dev/null
-
-# install
-cd ${ddpath} | ${bin} install -f -c ${cfg} -p p2
-
-cat ${tmpd}/abc
-grep '^this is some test' ${tmpd}/abc >/dev/null
-grep '^42' ${tmpd}/abc >/dev/null
-grep '^another test' ${tmpd}/abc >/dev/null
-
-#cat ${tmpd}/abc
+# checks
+[ ! -e ${tmpd}/abc ] && exit 1
 
 ## CLEANING
 rm -rf ${tmps} ${tmpd}
